+++ src/main/java/kitchenpos/dto/TableGroupResponse.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.order.dto.OrderTableResponse;
 
 public class TableGroupResponse {
