+++ src/main/java/kitchenpos/order/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 public class OrderTable {
