+++ src/main/java/kitchenpos/order/dto/OrderLineItemRequest.java
@@ -1,20 +1,24 @@
-package kitchenpos.domain;
+package kitchenpos.order.dto;
 
-public class OrderLineItem {
+import kitchenpos.order.domain.OrderLineItem;
+
+public class OrderLineItemRequest {
 
     private Long seq;
     private Long orderId;
     private Long menuId;
     private long quantity;
 
-    public OrderLineItem() {
+    public OrderLineItemRequest() {
     }
 
-    public OrderLineItem(Long seq, Long orderId, Long menuId, long quantity) {
-        this.seq = seq;
-        this.orderId = orderId;
+    public OrderLineItemRequest(Long menuId, long quantity) {
         this.menuId = menuId;
         this.quantity = quantity;
+    }
+
+    public OrderLineItem toOrderLineItem() {
+        return new OrderLineItem(seq, orderId, menuId, quantity);
     }
 
     public Long getSeq() {
