+++ src/main/java/kitchenpos/order/dto/OrderLineItemResponse.java
@@ -1,24 +1,31 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
-public class OrderLineItemRequest {
+public class OrderLineItemResponse {
 
     private Long seq;
     private Long orderId;
     private Long menuId;
     private long quantity;
 
-    public OrderLineItemRequest() {
+    public OrderLineItemResponse() {
     }
 
-    public OrderLineItemRequest(Long menuId, long quantity) {
+    public OrderLineItemResponse(Long seq, Long orderId, Long menuId, long quantity) {
+        this.seq = seq;
+        this.orderId = orderId;
         this.menuId = menuId;
         this.quantity = quantity;
     }
 
-    public OrderLineItem toOrderLineItem() {
-        return new OrderLineItem(seq, orderId, menuId, quantity);
+    public static OrderLineItemResponse from(OrderLineItem orderLineItem) {
+        return new OrderLineItemResponse(
+            orderLineItem.getSeq(),
+            orderLineItem.getOrderId(),
+            orderLineItem.getMenuId(),
+            orderLineItem.getQuantity()
+        );
     }
 
     public Long getSeq() {
