+++ src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -1,23 +1,32 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
 import java.time.LocalDateTime;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 
-public class OrderResponse {
+public class OrderRequest {
 
     private Long id;
     private Long orderTableId;
     private String orderStatus;
     private LocalDateTime orderedTime;
-    private List<OrderLineItemResponse> orderLineItems;
+    private List<OrderLineItemRequest> orderLineItems;
 
-    public OrderResponse() {
+    public OrderRequest() {
     }
 
-    public OrderResponse(Long id, Long orderTableId, String orderStatus,
-        LocalDateTime orderedTime, List<OrderLineItemResponse> orderLineItems) {
+    public OrderRequest(String orderStatus) {
+        this(null, null, orderStatus, null, Collections.emptyList());
+    }
+
+    public OrderRequest(Long orderTableId, List<OrderLineItemRequest> orderLineItems) {
+        this(null, orderTableId, null, null, orderLineItems);
+    }
+
+    public OrderRequest(Long id, Long orderTableId, String orderStatus,
+        LocalDateTime orderedTime, List<OrderLineItemRequest> orderLineItems) {
         this.id = id;
         this.orderTableId = orderTableId;
         this.orderStatus = orderStatus;
@@ -25,15 +34,14 @@ public class OrderResponse {
         this.orderLineItems = orderLineItems;
     }
 
-    public static OrderResponse from(Order order) {
-        return new OrderResponse(
-            order.getId(),
-            order.getOrderTableId(),
-            order.getOrderStatus(),
-            order.getOrderedTime(),
-            order.getOrderLineItems()
-                .stream()
-                .map(OrderLineItemResponse::from)
+    public Order toOrder() {
+        return new Order(
+            id,
+            orderTableId,
+            orderStatus,
+            orderedTime,
+            orderLineItems.stream()
+                .map(OrderLineItemRequest::toOrderLineItem)
                 .collect(Collectors.toList())
         );
     }
@@ -70,11 +78,11 @@ public class OrderResponse {
         this.orderedTime = orderedTime;
     }
 
-    public List<OrderLineItemResponse> getOrderLineItems() {
+    public List<OrderLineItemRequest> getOrderLineItems() {
         return orderLineItems;
     }
 
-    public void setOrderLineItems(final List<OrderLineItemResponse> orderLineItems) {
+    public void setOrderLineItems(final List<OrderLineItemRequest> orderLineItems) {
         this.orderLineItems = orderLineItems;
     }
 }
