+++ src/main/java/kitchenpos/order/dto/OrderTableRequest.java
@@ -1,30 +1,46 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
-public class OrderTableResponse {
+public class OrderTableRequest {
 
     private Long id;
     private Long tableGroupId;
     private int numberOfGuests;
     private boolean empty;
 
-    public OrderTableResponse() {
+    public OrderTableRequest() {
     }
 
-    public OrderTableResponse(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
+    public OrderTableRequest(Long id) {
+        this(id, null, 0, false);
+    }
+
+    public OrderTableRequest(int numberOfGuest) {
+        this(null, null, numberOfGuest, false);
+    }
+
+    public OrderTableRequest(boolean empty) {
+        this(null, null, 0, empty);
+    }
+
+    public OrderTableRequest(Long tableGroupId, int numberOfGuest, boolean empty) {
+        this(null, tableGroupId, numberOfGuest, empty);
+    }
+
+    public OrderTableRequest(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
         this.id = id;
         this.tableGroupId = tableGroupId;
         this.numberOfGuests = numberOfGuests;
         this.empty = empty;
     }
 
-    public static OrderTableResponse from(OrderTable orderTable) {
-        return new OrderTableResponse(
-            orderTable.getId(),
-            orderTable.getTableGroupId(),
-            orderTable.getNumberOfGuests(),
-            orderTable.isEmpty()
+    public OrderTable toOrderTable() {
+        return new OrderTable(
+            id,
+            tableGroupId,
+            numberOfGuests,
+            empty
         );
     }
